Add requests and responses logs
Requests and responses should be logged.

To achieve this, the getInputStream and getOutputStream is removed.
The getOutputStream is replaced by two send methods, one that takes
a byte array and one that takes a String. The only difference is how
the log output will look like.

The getInputStream is replaced with a Supplier, which is obtained
when the connect method is called.

The get method on the supplier waits until there is a response
available, or if there is already a response ready, it will return
that response immediately and clear the response, meaning calling
get multiple times will have a different effect.