Remove shells from properties
The shell manager can take a map of shells, which means that
the ShellImpl can be initialized from the factory class.

This means that the static create method can be removed in the ShellImpl.
Howerver, the class still needs to be public so that it can be initialized
from the framework.