Add custom server response functionality
Added a handler so that it's possible to define the server response.
Since the server can be used in a fixture it's only possible
to have one active response at the time.
Different responses will be queued.

Some of the interfaces lacks implementation and the
implementations added is not added to the factory.

This will be done in a following commit.